chore: use `getEnginesConfig` for template update linting
